Inline MerkleTree to OptimisedMerkleTree
